% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{adjustBudget}
\alias{adjustBudget}
\title{adjustBudget function}
\usage{
adjustBudget(bsp, targetBudget, targetStages)
}
\arguments{
\item{bsp}{A list of objects to combine with the species and population parameters. bsp is short for breeding sheme parameters}

\item{targetBudget}{Numeric value that you want the budget adjusted to}

\item{targetStages}{Character vector with stage names to be changed such that they become bigger or smaller to achieve the target budget}
}
\value{
A revised bsp with the sizes of the target stages changed to match.
}
\description{
Function to call once you have fully specified the costs but you want to adjust the size of the different stages so that the overall scheme budget matches some value.
}
\details{
Call this function after running specifyCosts.
}
\examples{
bsp <- adjustBudget(bsp, targetBudget=50000, targetStages=c("CET", "AYT", "UYT"))

}
