% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{adjustEntriesToBudget}
\alias{adjustEntriesToBudget}
\title{adjustEntriesToBudget function}
\usage{
adjustEntriesToBudget(
  bsp,
  targetBudget,
  fixedEntryStages = NULL,
  adjustStages = setdiff(bsp$stageNames, names(fixedEntryStages))
)
}
\arguments{
\item{bsp}{A list of objects to combine with the species and population parameters. bsp is short for breeding sheme parameters}

\item{targetBudget}{Numeric value that you want the budget adjusted to}

\item{fixedEntryStages}{Named integer vector indicating entry numbers for specific stages. Names must be names of the specific stages}

\item{adjustStages}{Character vector with names of stages to be changed such that the target budget is achieved}
}
\value{
A revised bsp with the sizes of the target stages changed to match.
}
\description{
Specify a budget, the number of entries for a set of stages, and the stages to adjust to hit the budget. The rules are that the first stage can't be bigger than the number of F1s, and no later stage can be bigger than an earlier stage. The function will not adjust if the rules are broken but will report.
}
\details{
Call this function after running specifyCosts.
}
\examples{
bsp <- adjustEntriesToBudget(bsp, targetBudget=50000, fixedEntryStages=c(PYT=100), adjustStages=c("CET", "AYT", "UYT"))

}
