% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{calculateBudget}
\alias{calculateBudget}
\title{calculateBudget function}
\usage{
calculateBudget(bsp)
}
\arguments{
\item{bsp}{A list of objects to combine with the species and population parameters. bsp is short for breeding sheme parameters}

\item{ctrlFileName}{The name of the text file with parameter values specifying the breeding costs. Must include the path to the file. If NULL a toy example simulation will be set up}
}
\value{
The bsp list with augmented or modified intermediate and total costs
}
\description{
Once the costs are specified in bsp, this function calculates the total annual budget for the breeding scheme
}
\details{
Call this function once costs have been specified
}
\examples{
bsp <- calculateBudget(bsp)

}
