% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{calculateChkReps}
\alias{calculateChkReps}
\title{calculateChkReps function}
\usage{
calculateChkReps(bsp)
}
\arguments{
\item{bsp}{A list of objects to combine with the species and population parameters. bsp is short for breeding sheme parameters}
}
\value{
The bsp list with updated chkReps
}
\description{
Once entries are specified, calculate the number of times checks will be replicated. Each rep must have at least one check.
}
\details{
Call this function once nEntries, nReps, nChks, and entryToChkRatio have been specified
}
\examples{
bsp <- calculateChkReps(bsp)

}
