% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFront.R
\name{checkAddRemove}
\alias{checkAddRemove}
\title{checkAddRemove function}
\usage{
checkAddRemove(cand, nonDom, var1name, var2name)
}
\arguments{
\item{cand}{Real double that you want to check if it is dominated or not}

\item{nonDom}{Real two-column matrix, with vectors in rows that are 
non-dominated and sorted from low to high for their first column}

\item{var1name}{Character: name of the column of column 1}

\item{var2name}{Character: name of the column of column 2}
}
\value{
A new nonDom matrix, adding cand if it is non-dominated and
removing any rows that are dominated by it
}
\description{
Test if any of the points in nonDom are dominated by pt
}
\details{
Utility for Pareto front
}
