% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{framePhenoRec}
\alias{framePhenoRec}
\title{framePhenoRec function}
\usage{
framePhenoRec(records)
}
\arguments{
\item{records}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
