% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{gebvPhenoEval}
\alias{gebvPhenoEval}
\title{Predict the GEBV of all genotyped individuals}
\usage{
gebvPhenoEval(phenoDF, grm)
}
\arguments{
\item{phenoDF}{A data.frame of phenotypic observations. See \code{framePhenoRec} for details}

\item{grm}{A genomic relationship matrix}
}
\value{
Named real vector of the BLUPs of all individuals in phenoDF (names are the individual ids), with appropriate weights by error variance of the observation
}
\description{
Function to predict the GEBV of all genotyped individuals in the population using all available phenotypes.
Takes a data.frame coming from \code{framePhenoRec} and
an additive kinship matrix (GRM) and analyze them with individuals .
Uses \code{\link[sommer]{mmer}} to implement a genomic BLUP model.
}
\details{
Given all the phenotypic records calculate the GEBV for each individual using all its records
}
