% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{genomicMateSelCrit}
\alias{genomicMateSelCrit}
\title{Generate predictions of cross means or usefulness as genomic mate selection criteria}
\usage{
genomicMateSelCrit(records, candidates, trainingpop, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{candidates}{Character vector of ids of the candidates to be parents, not necessarily phenotyped but must be predicted}

\item{trainingpop}{chr. vector of ids with phenotypes , but not necessarily in the list of selection candidates, but who should be included in the grm / training model.}

\item{bsp}{The breeding scheme parameter list}

\item{SP}{The AlphaSimR SimParam object (needed to pull SNPs)}
}
\value{
Character vector of the ids of the selected individuals
}
\description{
Mates are selected by these criteria e.g. by the \code{popImprovByMateSel populationImprovement function}
Additional arguments to add to the \code{bsp} object to accomplish this:
}
\details{
\itemize{
 \item \code{modelType}:
 \item \code{crossSelCrit}:
 \item \code{propSel}:
}
Setting up to distinguish between parent and cross selection and additive and non-additive predictions.
Uses \code{genomicMateSelectR} functions, which will need to be installed for these to work.

Modified original \code{selCritIID} function.

Accesses all individuals in \code{records} to pick the highest ones
}
