% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeByLOESS.R
\name{getParmsResponse}
\alias{getParmsResponse}
\title{getParmsResponse function}
\usage{
getParmsResponse(oneSim, startCycle)
}
\arguments{
\item{oneSim}{Output of runBreedingScheme}

\item{startCycle}{Integer first cycle after burn-in is done}
}
\description{
Utility function for optimizeByLOESS
Pull interesting parameters from from the stageOutputs that will serve to calculate optimality of the simulation. Currently the mean and genetic std dev of the F1 population
}
\details{
Takes a simulation output and returns a vector
}
