% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{grmPhenoEval}
\alias{grmPhenoEval}
\title{grmPhenoEval function}
\usage{
grmPhenoEval(phenoDF, grm)
}
\arguments{
\item{phenoDF}{A data.frame of phenotypic observations. See \code{framePhenoRec} for details}

\item{grm}{A genomic relationship matrix}
}
\value{
Named real vector of the BLUPs of all individuals in phenoDF (names are the individual ids), with appropriate weights by error variance of the observation
}
\description{
function to take a data.frame coming from framePhenoRec and GRM and analyze them with individuals as a random effect with a GRM covariance matrix
}
\details{
Given all the phenotypic records calculate the GEBV for each individual using all its records
}
\examples{
phenoDF <- framePhenoRec(records, bsp)
grm <- makeGRM(records, bsp, SP)
grmBLUPs <- grmPhenoEval(phenoDF, grm)

}
