% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{iidPhenoEval}
\alias{iidPhenoEval}
\title{iidPhenoEval function}
\usage{
iidPhenoEval(phenoDF)
}
\arguments{
\item{phenoDF}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
