% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFunctions.R
\name{initFuncADChk}
\alias{initFuncADChk}
\title{initFuncADChk function}
\usage{
initFuncADChk(bsp)
}
\arguments{
\item{bsp}{A list of breeding scheme parameters. It contains pipeline parameters: nParents, nCrosses, nProgeny, checks, nStages, errVars, nReps, nEntries, nChks, stageNames, and nCyclesToKeepRecords. It contains population parameters: nFounders, nChr, segSites, nQTL, genVar, meanDD, varDD, nSNP}
}
\value{
A list containing: 1. The simulation parameters in \code{SP}; 2. The initial records of the breeding program in \code{records}. See \code{fillPipeline} for details; 3. A completed \code{bsp} object
}
\description{
function to initialize simulation of a breeding program. A single additive-dominance trait is simulated.
}
\details{
Creates the founders and the initial records at the beginning of the simulation of a breeding program.
}
\examples{
none

}
