% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFunctions.R
\name{initializeFunc}
\alias{initializeFunc}
\title{initializeFunc function}
\usage{
initializeFunc(spp = NULL, ppp = NULL)
}
\arguments{
\item{spp}{A list of species parameters. It contains nFounders, nChr, segSites, nQTL, genVar, meanDD, varDD, nSNP.  See \code{runBreedingScheme} for details}

\item{ppp}{A list of product pipeline parameters.  It contains nParents, nCrosses, nProgeny, checks, nStages, errVars, nReps, nEntries, nChks, trialTypeNames, and nCyclesToKeepRecords.  See \code{runBreedingScheme} for details}
}
\value{
A list containing: 1. The simulation parameters in \code{SP}; 2. The initial records of the breeding program in \code{records}. See \code{fillPipeline} for details; 3. A list of product pipeline parameters containing nCrosses, nProgeny, checks, nStages, errVars, nReps, nEntries, nChks, trialTypeNames, and nCyclesToKeepRecords.
}
\description{
function to initialize simulation of a breeding program. A single additive-dominance trait is simulated.
}
\details{
Creates the founders and the initial records at the beginning of the simulation of a breeding program.
}
\examples{
initList <- initializeFunc()
SP <- initList$SP
records <- productPipeline(initList$records, ppp=initList$ppp)
records <- populationImprovement(initList$records, ppp=initList$ppp)
recordMeans <- mean_records(records)

}
