% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFunctions.R
\name{initializeScheme}
\alias{initializeScheme}
\title{initializeScheme function}
\usage{
initializeScheme(bsp)
}
\arguments{
\item{bsp}{A list of breeding scheme parameters.  See \code{specifyPipeline} and \code{specifyPopulation}}
}
\value{
A list containing: 1. The simulation parameters in \code{SP}; 2. The initial records of the breeding program in \code{records}. See \code{fillPipeline} for details; 3. A completed \code{bsp} object
}
\description{
function to initialize simulation of a breeding program. A single additive-dominance trait is simulated. Check are used in this scheme
}
\details{
Creates the founders and the initial records at the beginning of the simulation of a breeding program.
}
\examples{
bsp <- specifyPopulation(bsp)
bsp <- specifyPipeline()
initList <- initializeScheme(bsp)
SP <- initList$SP
bsp <- initList$bsp
records <- initList$records

}
