% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFunctions.R
\name{initializeSimParam}
\alias{initializeSimParam}
\title{initializeSimParam function}
\usage{
initializeSimParam(meanDD = 0.2, varDD = 0.1, nSNPperChr = 100)
}
\arguments{
\item{meanDD}{[value]. Default is meanDD=0.2}

\item{varDD}{[value]. Default is varDD=0.1}

\item{nSNPperChr}{[value]. Default is nSNPperChr=100}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
