% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFront.R
\name{insertRow}
\alias{insertRow}
\title{insertRow function}
\usage{
insertRow(vec, idx, mat)
}
\arguments{
\item{vec}{Real row vector}

\item{idx}{Integer insertion point}

\item{mat}{Real matrix}
}
\value{
Augmented matrix
}
\description{
Insert a vector (vec) into a matrix (mat) at row idx
}
\details{
Utility for Pareto front
}
