% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\docType{data}
\name{ip}
\alias{ip}
\title{packages installed at the machine}
\format{
An object of class \code{character} of length 270.
}
\usage{
ip
}
\description{
packages installed at the machine
}
\keyword{datasets}
