% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulnessSim.R
\name{kinshipHlpR}
\alias{kinshipHlpR}
\title{kinshipHlpR function}
\usage{
kinshipHlpR(records, bsp, SP, type, snpsORqtl = "snps")
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{The breeding scheme parameter list}

\item{SP}{The AlphaSimR SimParam object. Needed to pull the SNP genotypes}

\item{type}{string, "add" or "dom". type="add" gives same as rrBLUP::A.mat(), i.e. Van Raden, Method 1. type="dom" gives classical parameterization according to Vitezica et al. 2013.}

\item{snpsORqtl}{string, "snps" (default) or "qtl". Determines which loci are used to compute the GRM.}
}
\value{
A genomic relationship matrix
}
\description{
function to make either additive or dominance genomic relationship matrices using predCrossVar::kinship() at either SNPs or QTL.
}
\details{
\code{records} maintains the phenotypic and genotypic records across years and stages. For GEBV analysis, you need the GRM of these individuals. \code{makeGRM} assumes the first phenotyping stage (records[[2]]) has all individuals that have been phenotyped. The GRM also includes the unphenotyped new F1 individuals in records[[1]]
}
\examples{
grm <- makeGRM(records, bsp, SP)

}
