% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{lastCycStgOut}
\alias{lastCycStgOut}
\title{lastCycStgOut function}
\usage{
lastCycStgOut(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{A list of breeding scheme parameters}

\item{SP}{the AlphaSimR SimParam object}
}
\value{
A records object that has updated stageOutputs
}
\description{
stageOutputs are always made on the previous cycle, so to get the end, you need to run this once
}
\details{
Will figure this out later...
}
\examples{
records <- lastCycStgOut(records, bsp, SP)
}
