% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{make_grm}
\alias{make_grm}
\title{Function to make genomic relation matrices}
\usage{
make_grm(records, bsp, SP, grmType = "add")
}
\arguments{
\item{records}{The breeding program \code{records} object. See
\code{fillPipeline} for details}

\item{bsp}{The breeding scheme parameter list}

\item{SP}{The AlphaSimR SimParam object. Needed to pull the SNP genotypes}
}
\value{
A genomic relationship matrix
}
\description{
Function to make a genomic relationship matrix to be used to analyze the
phenotypic \code{records}
}
\details{
Setting up to distinguish between parent and cross selection and additive and non-additive predictions.
Uses \code{genomicMateSelectR} functions, which will need to be installed for these to work.

This version uses the \code{genomicMateSelectR::kinship}.

Includes a new \code{grmType} param, can produce additive and dominance matrices.
Same matrix as \code{A.mat} by default \code{grmType="add"}.

\code{records} maintains the phenotypic and genotypic records across
  years and stages. For GEBV analysis, you need the GRM of these individuals.
  \code{makeGRM} assumes the first phenotyping stage (records[[2]]) has all
  individuals that have been phenotyped. The GRM also includes the
  unphenotyped new F1 individuals in records[[1]]
}
