% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{mean_records}
\alias{mean_records}
\title{mean_records function}
\usage{
mean_records(records)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}
}
\value{
Real matrix with breeding program cycles in rows and product pipeline stages in columns, each cell being the mean genotypic value for that year and stage
}
\description{
function to calculate the mean genotypic value at each cycle and stage
WARNING: right now, checks are kept in the population records and they are included in the means that are calculated
}
\details{
The records object is a list of lists of populations. This function takes those lists and returns the poplation means in a matrix
}
\examples{
recordMeans <- mean_records(records)

}
