% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{optContrib}
\alias{optContrib}
\title{optContrib function}
\usage{
optContrib(records, bsp, SP, crit)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{A list of product pipeline parameters}

\item{SP}{The AlphaSimR SimParam object (needed to pull SNPs)}

\item{crit}{Named vector of selection criterion to be maximized}
}
\value{
Pop class object with the progeny from optimum contribution crosses
}
\description{
function uses optiSel to identify number of progeny, allocate mates to minimize inbreeding depression, and return progeny
}
\details{
Calculate a grm of individuals with high enough crit, then maximize crit subject to a target increase of relatedness consistent with bsp$targetEffPopSize
}
\examples{
crit <- bv(records$F1); names(crit) <- records$F1@id
progeny <- optContrib(records, bsp, SP, crit)
}
