% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{parentSelCritGEBV}
\alias{parentSelCritGEBV}
\title{parentSelCritGEBV function}
\usage{
parentSelCritGEBV(records, candidates, trainingpop, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{candidates}{Character vector of ids of the candidates to be parents, not necessarily phenotyped but must be predicted}

\item{trainingpop}{chr. vector of ids with phenotypes , but not necessarily in the list of selection candidates, but who should be included in the grm / training model.}

\item{bsp}{The breeding scheme parameter list}

\item{SP}{The AlphaSimR SimParam object (needed to pull SNPs)}
}
\value{
Character vector of the ids of the selected individuals
}
\description{
\code{parentSelCritGEBV} will compute GEBV of all genotyped individuals using all available phenotypes.
Parents are selected by these criteria e.g. by the \code{popImprov1Cyc populationImprovement function}
}
\details{
Setting up to distinguish between parent and cross selection and additive and non-additive predictions.
Uses \code{genomicMateSelectR} functions, which will need to be installed for these to work.

Modified original \code{selCritGRM} function.
Uses alternative functions \code{make_grm} and \code{gebvPhenoEval}.

Accesses all individuals in \code{records} to pick the highest ones
}
\examples{

}
