% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{phenoRecFromPop}
\alias{phenoRecFromPop}
\title{phenoRecFromPop function}
\usage{
phenoRecFromPop(pop, bsp, stage, checks = FALSE)
}
\arguments{
\item{pop}{The population from which to extract phenotypic records. Has to have been phenotyped}

\item{bsp}{The breeding scheme parameter list}

\item{stage}{At what stage of the breeding scheme this population was phenotyped. Necessary to determine the error variance and degree of replication}

\item{checks}{Whether this was a population of experimentals or checks. Necessary to determine the degree of replication}
}
\value{
A tibble of phenotypic records with six columns: 1. The id, 2. The id of the mother, 3. The id of the father, 4. The name of the stage, 5. The actual phenotype, 6. The error variance of the phenotype
}
\description{
function to make a tibble to be added to \code{records}
}
\details{
The tibbles coming from this function will be incorporated into \code{records} useful for maintaining the phenotypic observations across years and stages. For analysis, you need these phenotypes coupled to their ids and error variances
}
\examples{
phenoDF <- phenoRecFromPop(pop, bsp, stage)

}
