% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotRecords}
\alias{plotRecords}
\title{plotRecords function}
\usage{
plotRecords(replicRecords)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}
}
\value{
Makes a plot of the gain over cycles of selection and returns the dataset used to make that plot
}
\description{
function to make a plot from the data in \code{records}
}
\details{
\code{records} is a list of lists of populations and phenotype matrices useful for maintaining the phenotypic observations across years and stages
}
\examples{
exptSummary <- plotRecords(records)

}
