% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{popImprov1}
\alias{popImprov1}
\title{popImprov1 function}
\usage{
popImprov1(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{A list of breeding scheme parameters}

\item{SP}{The AlphaSimR SimParam object}

\item{selectFunc}{Function to select parents among candidates. Default is the AlfSimHlpR function selectParIID}
}
\value{
A records object with a new F1 Pop-class object of progeny coming out of a population improvement scheme
}
\description{
Function to improve a simulated breeding population by one cycle. This version takes phenotyped individuals and crosses them to create new F1
}
\details{
This function uses penotypic records coming out of the product pipeline to choose individuals as parents to initiate the next breeding cycle
}
\examples{
none
}
