% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{popImprov2Cyc}
\alias{popImprov2Cyc}
\title{popImprov2Cyc function}
\usage{
popImprov2Cyc(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{List of breeding scheme parameters}

\item{SP}{The AlphaSimR SimParam object}
}
\value{
A records object with a new F1 Pop-class object of progeny coming out of a population improvement scheme
}
\description{
Function to improve a simulated breeding population by one cycle. This version does two cycles of predicting F1 individuals and making new F1s
}
\details{
This function uses penotypic records coming out of the product pipeline to choose individuals as parents to initiate the next breeding cycle
}
\examples{
bsp <- specifyPipeline()
bsp <- specifyPopulation(bsp)
initList <- initializeFunc(bsp)
SP <- initList$SP
bsp <- initList$bsp
records <- initList$records
records <- prodPipeSimp(records, bsp, SP)
records <- popImprov2(records, bsp, SP)

}
