% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{popImprovByMateSel}
\alias{popImprovByMateSel}
\title{Run population improvement using mate selection}
\usage{
popImprovByMateSel(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{A list of breeding scheme parameters}

\item{SP}{The AlphaSimR SimParam object}
}
\value{
A records object with a new F1 Pop-class object of progeny coming out of a population improvement scheme
}
\description{
Function to improve a simulated breeding population by one cycle.
Also specify \code{selCritPop="genomicMateSelCrit"} and use the following new \code{bsp} arguments to control the predictions and selection:
\itemize{
 \item \code{modelType}: modelTypes: "A", "AD", "DirDom"
 \item \code{crossSelCrit}: "MeanBV" (modelTypes: "A", "AD", "DirDom"); "MeanTGV" (only with modelType="DirDom"); "UCparent" (modelTypes: "A", "AD", "DirDom"); "UCvariety" (modelTypes: "AD", "DirDom")
 \item \code{propSel}: 0-1, prop. of predicted crosses to select, used ONLY to calc the standardized selection intensity (i) and subsequently the usefulness criteria (\eqn{\mu + i \times \sigma}).
}
}
\details{
My changes:
\itemize{
 \item \code{nTrainPopCycles}: draw training pop clones only from this number of recent cycles.
 \item \code{nYrsAsCandidates}: candidates for selection only from this number of recent years
 \item \code{maxTrainingPopSize}: From the lines in the most recent cycles (indicated by \code{nTrainPopCycles}),
 subsample this number of lines for training data. This is \emph{in addition to} the "check" (\code{bsp$checks@id})
 and the lines indicates as selection `candidates` according to the setting of `nYrsAsCandidates`.
 All "historical" data will always be used, but the number of maximum training lines will be held constant.
 Replaces the stage-specific `bsp$trainingPopCycles`, which will be unused in this pipeline, but not deleted from the package.
 }

This function uses phenotypic records coming out of the product pipeline to choose individuals as parents to initiate the next breeding cycle
}
