% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{popImprovByParentSel}
\alias{popImprovByParentSel}
\title{Run population improvement using parent selection}
\usage{
popImprovByParentSel(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{A list of breeding scheme parameters}

\item{SP}{The AlphaSimR SimParam object}
}
\value{
A records object with a new F1 Pop-class object of progeny coming out of a population improvement scheme
}
\description{
Function to improve a simulated breeding population by one cycle.
This version is adjusted relative to the original \code{popImprov1Cyc} function,
which drew \code{candidates} from the full \code{records$F1} (excluding potentially
indivs only scored during the current year, if \code{useCurrentPhenoTrain=FALSE}).
My changes:
\itemize{
 \item \code{nTrainPopCycles}: draw training pop clones only from this number of recent cycles.
 \item \code{nYrsAsCandidates}: candidates for selection only from this number of recent years
 \item \code{maxTrainingPopSize}: From the lines in the most recent cycles (indicated by \code{nTrainPopCycles}),
 subsample this number of lines for training data. This is \emph{in addition to} the "check" (\code{bsp$checks@id})
 and the lines indicates as selection `candidates` according to the setting of `nYrsAsCandidates`.
 All "historical" data will always be used, but the number of maximum training lines will be held constant.
 Replaces the stage-specific `bsp$trainingPopCycles`, which will be unused in this pipeline, but not deleted from the package.
 }
}
\details{
This function uses penotypic records coming out of the product pipeline to choose individuals as parents to initiate the next breeding cycle
}
