% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{popImprovOutput}
\alias{popImprovOutput}
\title{popImprovOutput function}
\usage{
popImprovOutput(records, crit, candidates, trainingpop, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{crit}{Named vector of selection criterion to be maximized}

\item{candidates}{Character vector of ids of the candidates to be parents}

\item{trainingpop}{Character vector of ids of the training population selected to predict the candidates}

\item{SP}{The AlphaSimR SimParam object}
}
\value{
A tibble with whatever information from the parent selection to improve the breeding population you want to store for analysis after simulation is done
}
\description{
To create a row for a tibble for each cycle of population improvement. The tibble will be kept as the last object of the records list
}
