% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulnessSim.R
\name{popImprovUC}
\alias{popImprovUC}
\title{popImprovUC function}
\usage{
popImprovUC(records, bsp, SP, cycle)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{A list of breeding scheme parameters}

\item{SP}{The AlphaSimR SimParam object}

\item{cycle}{An integer for the current breeding cycle. This slot was added so that it could be passed from the runBreedingSchemeCycleOutputs function for tracking purposes.}
}
\value{
A records object with a new F1 Pop-class object of progeny coming out of a population improvement scheme
}
\description{
Version of the popImprov1cyc that runs selCritUCA function to improve a simulated breeding population by one cycle.
This version obtains additive Usefulness Criterion (UC) for crosses and creates evenly sized new families from the best crosses.
This version also stores the \code{records$cycleOutputs} including e.g. varcomps and SNP effects
}
\details{
This function uses penotypic records coming out of the product pipeline to choose individuals as parents to initiate the next breeding cycle
}
\examples{
bsp <- specifyPipeline()
bsp <- specifyPopulation(bsp)
initList <- initializeFunc(bsp)
SP <- initList$SP
bsp <- initList$bsp
records <- initList$records
records <- prodPipeSimp(records, bsp, SP)
records <- popImprovUC(records, bsp, SP)

}
