% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{populationImprovement}
\alias{populationImprovement}
\title{populationImprovement function}
\usage{
populationImprovement(records, ppp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{ppp}{A list of product pipeline parameters.  It contains nParents, nCrosses, nProgeny, checks, nStages, errVars, nReps, nEntries, nChks, trialTypeNames, and nCyclesToKeepRecords.  See \code{initializeFunc} for details}
}
\value{
A records object with a new F1 Pop-class object of progeny coming out of a population improvement scheme
}
\description{
function to improve a simulated breeding population by one cycle. See Gaynor et al. 2017 for the general idea.
}
\details{
This function uses penotypic records coming out of the product pipeline to choose individuals as parents to initiate the next breeding cycle
}
\examples{
initList <- initializeFunc()
SP <- initList$SP
records <- productPipeline(initList$records, ppp=initList$ppp, selectFunc=selectAdvance)
records <- populationImprovement(records)
recordMeans <- mean_records(records)
}
