% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{prodPipeSimp}
\alias{prodPipeSimp}
\title{prodPipeSimp function}
\usage{
prodPipeSimp(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{A list of breeding scheme parameters}

\item{SP}{the AlphaSimR SimParam object}

\item{selectFunc}{In this case a dummy function. \code{runBreedingScheme} passes it but this simple function doesn't use it}
}
\value{
A records object that has new records created by advancing by a generation
}
\description{
Simple function to advance a simulated breeding product pipeline forward by one generation. No use-input function, no checks. Selection on phenotype
}
\details{
The breeding program product pipeline will have been set by initializeFunc. This function moves the breeding program along by one generation and saves all the resulting phenotypes to the records object.
}
\examples{
none

}
