% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{productPipeline}
\alias{productPipeline}
\title{productPipeline function}
\usage{
productPipeline(records, checks = NULL, selectFunc = NULL)
}
\arguments{
\item{records}{[value]}

\item{checks}{[value]. Default is checks=NULL}

\item{selectFunc}{[value]. Default is selectFunc=NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
