% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProductPipeline.R
\name{productPipelinePostBurnIn}
\alias{productPipelinePostBurnIn}
\title{productPipeline function for when a new bsp object has been supplied}
\usage{
productPipelinePostBurnIn(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{the original parameter set used}

\item{SP}{the AlphaSimR SimParam object}

\item{newbsp}{the new list of breeding scheme parameters}
}
\value{
A records object that has new records created by advancing by a generation
}
\description{
Example: post burn-in
}
\details{
function to advance a simulated breeding product pipeline forward by one generation. See Gaynor et al. 2017 for the general idea.

The breeding program product pipeline will have been set by initializeFunc. This function moves the breeding program along by one generation and saves all the resulting phenotypes to the records object.
}
\examples{
bsp <- specifyPipeline()
bsp <- specifyPopulation(bsp)
initList <- initializeFunc(bsp)
SP <- initList$SP
bsp <- initList$bsp
records <- initList$records
records <- productPipeline(records, bsp, SP)
records <- popImprov1(records, bsp, SP)

}
