% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{readControlFile}
\alias{readControlFile}
\title{function to read a text control file}
\usage{
readControlFile(fileName, parmNames)
}
\arguments{
\item{fileName}{The name of the text file to be read. Must include the path to the file}

\item{parmNames}{A string vector with the names of the control parameters that will be searched in the text file}
}
\value{
A named list of the parameter values read from the control file
}
\description{
The text file should be organized as follows
1. Any text after a comment symbol # will be ignored
2. Control parameter names should be on their own line
3. Parameter values should be on the following line. If multiple parameter values are needed they should be separated by white space but on the same line
}
\details{
Call this function before beginning the simulation
}
\examples{
params <- readControlFile("./inputDir/ctrlFile.txt", c("nStages", "nParents", "nCrosses"))

}
