% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyingFunctions.R
\name{removeOldestCyc}
\alias{removeOldestCyc}
\title{removeOldestCyc function}
\usage{
removeOldestCyc(records)
}
\arguments{
\item{records}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
