% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeByLOESS.R
\name{repeatSim}
\alias{repeatSim}
\title{repeatSim function}
\usage{
repeatSim(
  parmRow,
  replication,
  radius = 0.02,
  initializeFunc,
  productPipeline,
  populationImprovement,
  targetBudget,
  bsp,
  seed = NULL
)
}
\arguments{
\item{parmRow}{Vector resulting from getParmsResponse and that contains parameters for the simulation to be repeated}

\item{replication}{Integer replication}

\item{radius}{Real radius around which to sample the budget so it's not an _exact_ repeat}

\item{initializeFunc}{Function for runBreedingScheme}

\item{productPipeline}{Function for runBreedingScheme}

\item{populationImprovement}{Function for runBreedingScheme}

\item{targetBudget}{Real how much to spend total. See CostControlFile}

\item{bsp}{List breeding scheme parameters}

\item{seed=NULL}{Integer seed to initialize pseudo-random number generator}
}
\description{
Utility function for optimizeByLOESS
Run a simulation in the vicinity of a previous simulation
}
\details{
Sample the budget and run a simulation close to a previous one. Useful to rerun where the prediction is particularly high (we need to know if that prediction is for real) or where the standard error is particularly high (we need more info in that vicinity)
}
