% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{runBreedingScheme_wBurnIn}
\alias{runBreedingScheme_wBurnIn}
\title{Run a breeding scheme including a burn-in period}
\usage{
runBreedingScheme_wBurnIn(
  replication = NULL,
  bsp,
  nBurnInCycles,
  nPostBurnInCycles,
  selCritPopPre,
  selCritPopPost,
  selCritPipePre = "selCritIID",
  selCritPipePost = "selCritIID",
  initializeFunc,
  productPipeline,
  populationImprovement,
  nBLASthreads = NULL,
  nThreadsMacs2 = NULL
)
}
\arguments{
\item{replication}{Integer replication of running the breeding scheme}

\item{bsp}{A list of breeding scheme parameters.}

\item{nBurnInCycles}{Integer number of cycles to as 'burn-in' using the \code{selCritPopPre} and \code{selCritPipePre} settings.}

\item{nPostBurnInCycles}{Integer number of cycles to as 'burn-in' using the \code{selCritPopPost} and \code{selCritPipePost} settings.}

\item{selCritPopPre}{string, overrides the selCrit in \code{bsp} for the burn-in stage.}

\item{selCritPopPost}{string, overrides the selCrit in \code{bsp} for the burn-in stage.}

\item{selCritPipePre}{string, overrides the selCrit in \code{bsp} for the burn-in stage. DEFAULT: selCritIID}

\item{selCritPipePost}{string, overrides the selCrit in \code{bsp} for the burn-in stage. DEFAULT: selCritIID}

\item{initializeFunc}{Function to initialize the breeding program.}

\item{productPipeline}{Function to advance the product pipeline by one generation}

\item{populationImprovement}{Function to improve the breeding population and select parents to initiate the next cycle of the breeding scheme}

\item{nBLASthreads}{number of cores for each worker to use for multi-thread BLAS. Will speed up, for example, genomic predictions when using selCritGRM. Careful to balance with other forms of parallel processing.}

\item{nThreadsMacs2}{uses the nThreads argument in \code{runMacs2}, parallelizes founder sim by chrom.}
}
\value{
A \code{records} object containing the phenotypic records retained of the breeding scheme
}
\description{
Allows users to switch selection critera after a designated set of cycles and continue for additional cycles.
Use case envisioned is to run phenotypic selection for an initial period and then switch to e.g. GS.
Adds a few other bells and whistles to the original \code{runBreedingScheme} function.
}
\details{
A wrapper to initiate the breeding program then iterate cycles of product pipeline and population improvement
}
