% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFunctions.R
\name{runSchemesPostBurnIn}
\alias{runSchemesPostBurnIn}
\title{Continue running breeding schemes post burn-in}
\usage{
runSchemesPostBurnIn(
  simulations,
  newBSP = NULL,
  nPostBurnInCycles,
  selCritPop = "parentSelCritBLUP",
  selCritPipe = "productSelCritBLUP",
  productFunc = "productPipelinePostBurnIn",
  popImprovFunc = "popImprovByParentSel",
  nSimCores = 1,
  nBLASthreads = NULL
)
}
\arguments{
\item{simulations}{tibble, each row is a simulation, 2 columns: SimRep and burnInSim.
SimRep is an identifier. burnInSim is a list with 3 named elements:
"records", "bsp" and "SP"}

\item{newBSP}{optional, so you can specify a different bsp for post-burn in sims.}

\item{nPostBurnInCycles}{Integer number of cycles to run the \code{selCritPop} and \code{selCritPipe} settings.}

\item{selCritPop}{string, overrides the selCrit in \code{bsp} for the post burn-in stage.}

\item{selCritPipe}{string, overrides the selCrit in \code{bsp} for the post burn-in stage.}

\item{productFunc}{string, Function to advance the product pipeline by one generation}

\item{popImprovFunc}{string, Function to improve the breeding population and select parents to initiate the next cycle of the breeding scheme}

\item{nSimCores}{Integer, number of cores to optionally execute replicate simulations in parallel}

\item{nBLASthreads}{number of cores for each worker to use for multi-thread BLAS. Will speed up, for example, genomic predictions when using selCritGRM. Careful to balance with other forms of parallel processing.}
}
\value{
A \code{records} object containing the phenotypic records retained of the breeding scheme
}
\description{
Allows users to continue a simulation, potentially initiated or 'burned-in' with \code{runBurnInSchemes}.
Allows user to optionally change the \code{bsp} and selection criterion.
Input \code{simulations} are the output of e.g. \code{runBurnInSchemes}.
Runs potentially multiple replications and optionally in parallel.
}
\details{
A wrapper to initiate the breeding program then iterate cycles of product pipeline and population improvement
}
