% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{selCritIID}
\alias{selCritIID}
\title{selCritIID function}
\usage{
selCritIID(records, candidates, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{candidates}{Character vector of ids of the candidates to be parents}

\item{bsp}{The breeding scheme parameter list}

\item{SP}{The AlphaSimR SimParam object (not used, here for uniformity)}
}
\value{
An IID BLUP of the trait of the candidates
}
\description{
function to select parents among individuals with phenotypes, assuming individual effects are IID
}
\details{
Accesses all individuals in \code{records} to pick the highest among candidates. If candidates do not have records, a random sample is returned
}
\examples{
allPop <- mergePops(records[[2]])
candidates <- allPop@id
parents <- allPop[selCritIID(records, candidates, bsp, SP)]

}
