% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulnessSim.R
\name{selCritUCA}
\alias{selCritUCA}
\title{selCritUCA function}
\usage{
selCritUCA(
  records,
  candidates,
  bsp,
  SP,
  snpsORqtl = "snps",
  nPreselect = bsp$nParents,
  ncores = 1,
  propSel = 0.1,
  h = 1
)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{candidates}{Character vector of ids of the candidates to be parents}

\item{bsp}{The breeding scheme parameter list}

\item{SP}{The AlphaSimR SimParam object (needed to pull SNPs)}
}
\value{
tibble with 2 list-columns (selCrit: contains _per cross_ selection criterion, geneticEval: contains varcomps, marker effects and breeding values from GBLUP.)
}
\description{
function to compute the usefulness cross-selection criteria based on combo of parental GEBV and predicted _additive_ genetic variance of each cross.
}
\details{
Accesses all individuals in \code{records} to pick the highest ones
}
\examples{
candidates <- records[[1]][[1]]@id
parents <- records[[1]][[1]][selCritGRM(records, candidates, bsp, SP)]

}
