% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{selectAdvance}
\alias{selectAdvance}
\title{selectAdvance function}
\usage{
selectAdvance(popPheno, records, checks, ids)
}
\arguments{
\item{popPheno}{[value]}

\item{records}{[value]}

\item{checks}{[value]}

\item{ids}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
