% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{selectParents}
\alias{selectParents}
\title{selectParents function}
\usage{
selectParents(records, nToSelect)
}
\arguments{
\item{records}{[value]}

\item{nToSelect}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
