% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{selectParGRM}
\alias{selectParGRM}
\title{selectParGRM function}
\usage{
selectParGRM(records, candidates, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{candidates}{Character vector of ids of the candidates to be parents}

\item{bsp}{A list of product pipeline parameters}
}
\value{
Character vector of the ids of the selected individuals
}
\description{
function to select parents among individuals with phenotypes, assuming individual effects covary according to a GRM
}
\details{
Accesses all individuals in \code{records} to pick the highest ones
}
\examples{
none
}
