% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{selectParIID}
\alias{selectParIID}
\title{selectParIID function}
\usage{
selectParIID(records, candidates, bsp)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{candidates}{Character vector of ids of the candidates to be parents}

\item{bsp}{A list of product pipeline parameters}
}
\value{
Character vector of the ids of the selected individuals
}
\description{
function to select parents among individuals with phenotypes, assuming individual effects are IID
}
\details{
Accesses all individuals in \code{records} to pick the highest among candidates. If candidates do not have records, a random sample is returned
}
\examples{
none
}
