% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulnessSim.R
\name{snpeffPhenoEvalA}
\alias{snpeffPhenoEvalA}
\title{snpeffPhenoEvalA function}
\usage{
snpeffPhenoEvalA(phenoDF, grmA, zA)
}
\arguments{
\item{phenoDF}{A data.frame of phenotypic observations. See \code{framePhenoRec} for details}

\item{grmA}{The _additive_ genomic relationship matrix}

\item{zA}{Centered _additive_ marker dosage matrix of use as predictors when backsolving marker effects from GBLUP solutions}
}
\value{
tibble with slots for variance components, marker effects and BLUPs
}
\description{
function to take a data.frame coming from framePhenoRec and GRM and analyze them with individuals as a random effect with a GRM covariance matrix
}
\details{
Given all the phenotypic records calculate the GEBV and backsolve the SNP effects.
_Also_ computes the genomic variance according to Lehermeier et al. 2017a method "M2" (accounts for LD).
}
\examples{
phenoDF <- framePhenoRec(records, bsp)
grm <- makeGRM(records, bsp, SP)
grmBLUPs <- grmPhenoEval(phenoDF, grm)

}
