% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipelineChkRep.R
\name{specifyPipeline}
\alias{specifyPipeline}
\title{specifyPipeline function}
\usage{
specifyPipeline(bsp = NULL, ctrlFileName = NULL)
}
\arguments{
\item{bsp}{A list of objects to combine with the pipeline parameters. bsp is short for breeding sheme parameters}

\item{ctrlFileName}{The name of the text file with parameter values controling the simulation. Must include the path to the file. If NULL a toy example simulation will be set up}
}
\value{
A list containing objects that specify the product pipeline. This list will determine the number of lists in the records object
}
\description{
function specify the product pipeline. This would not need to be a function, but this way all definitions are in one place
}
\details{
Call this function before beginning the simulation
}
\examples{
bsp <- specifyPipeline()

}
