% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipelineChkRep.R
\name{specifyPopulation}
\alias{specifyPopulation}
\title{specifyPopulation function}
\usage{
specifyPopulation(bsp = NULL, ctrlFileName = NULL)
}
\arguments{
\item{bsp}{A list of objects to combine with the species and population parameters. bsp is short for breeding sheme parameters}

\item{ctrlFileName}{The name of the text file with parameter values specifying the breeding population. Must include the path to the file. If NULL a toy example simulation will be set up}
}
\value{
A list containing objects that specify the species and population characteristics.
}
\description{
Function to specify the species and population characteristics. This would not need to be a function, but this way all definitions are in one place
}
\details{
Call this function before beginning the simulation
}
\examples{
bsp <- specifyPopulation(bsp)

}
