% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{stageOutputs}
\alias{stageOutputs}
\title{stageOutputs function}
\usage{
stageOutputs(id, f1, selCrit, stage, year)
}
\arguments{
\item{id}{Vector of the AlphaSimR ids of the individuals in the stage}

\item{f1}{The AlphaSimR pop class with all the individuals}

\item{selCrit}{Named vector of the selection criterion being used to advance individuals}

\item{stage}{Integer stage (1 to bsp$nStages) being summarized}

\item{year}{The current year of the breeding scheme}
}
\value{
A tibble with whatever information from the data you want to store for analysis after simulation is done
}
\description{
To create a row for a tibble for each cycle and each stage. The tibble will be kept as the last object of the records list
}
\details{
Trying to provide some flexibility in what results AlphaSimHlpR generates from a given simulation.
}
\examples{
records$stageOutputs <- records$stageOutputs \%>\% bind_rows(stageOutputs(id, records$F1, selCrit, stage, year))

}
