% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFront.R
\name{testDom}
\alias{testDom}
\title{testDom function}
\usage{
testDom(pt, nonDom)
}
\arguments{
\item{pt}{Real double}

\item{nonDom}{Real two-column matrix}
}
\value{
Logical vector size number of matrix rows
}
\description{
Test if any of the points in nonDom are dominated by pt
}
\details{
Utility for Pareto front
}
