% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyingFunctions.R
\name{tidyrecords}
\alias{tidyrecords}
\title{tidyrecords function}
\usage{
tidyrecords(records)
}
\arguments{
\item{records}{a list-class "records" object produced by AlphaSimHlpR}
}
\value{
a tibble with indicator for year and stage
}
\description{
function to create a tibble with indicator for year and stage from a "records" object produced by AlphaSimHlpR. Possibly redundant to framePhenoRec() function.
}
\examples{
records <- fillPipeline(founders, bsp, SP)
records <- tidyrecords(records)
}
