% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyingFunctions.R
\name{tidysims}
\alias{tidysims}
\title{tidysims function}
\usage{
tidysims(sims)
}
\arguments{
\item{sims}{a list of AlphaSimHlpR sims (result of runBreedingScheme)}
}
\value{
a tidy tibble with one row per simulation, cols: records (tibble of all phenotypic records with indicator added for "year" and "stageName"), simulatedpop (the pop-class object from the sim), bsp and SP for the sim.
}
\description{
function to create a tidy tibble from a list of runBreedingScheme simulations
}
\examples{
sims <- map(1:2,~runBreedingScheme(replication = .,nCycles = 7,
                                initializeFunc = initFuncADChk,
                                productPipeline = prodPipeFncChk,
                                populationImprovement = popImprov1Cyc,
                                bsp = bsp))
sims <- tidysims(sims)
}
