% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{trainPopSelHRep}
\alias{trainPopSelHRep}
\title{trainPopSelHRep function}
\usage{
trainPopSelHRep(phenotypedLines_notSelCands, maxTPsize)
}
\arguments{
\item{phenotypedLines_notSelCands}{List of individuals available to compose the training population
of a genomic prediction, excluding the candidates for parent selection of the next cycle}

\item{maxTPSize}{Number of individuals to select to compose the training population of the genomic prediction}
}
\value{
Return a list of individuals that will be the training population of the genomic prediction
}
\description{
function select the individuals with higher number of repetitions to compose the training population of the genomic prediction
}
