% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{trainPopSelNew}
\alias{trainPopSelNew}
\title{trainPopSelNew function}
\usage{
trainPopSelNew(phenotypedLines_notSelCands, maxTPsize)
}
\arguments{
\item{phenotypedLines_notSelCands}{List of individuals available to compose the training population
of a genomic prediction, excluding the candidates for parent selection of the next cycle}

\item{maxTPSize}{Number of individuals to select to compose the training population of the genomic prediction}
}
\value{
Return a list of individuals that will be the training population of the genomic prediction
}
\description{
function select the last individuals created by the simulation to compose the training population of the genomic prediction
}
