% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{calcDerivedParms}
\alias{calcDerivedParms}
\title{calcDerivedParms function}
\usage{
calcDerivedParms(bsp)
}
\arguments{
\item{bsp}{A list. bsp is short for breeding sheme parameters.}
}
\value{
A list bsp that extends the input with a few derived parameters
}
\description{
Once you have read in parameters from a control file, or set them yourself, there are still a few derived parameters that are needed.  This function calculates them.
}
\details{
This function is only called internally by other functions used to specify the pipeline
}
