% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{fillPipeline}
\alias{fillPipeline}
\title{fillPipeline function}
\usage{
fillPipeline(parents, checks = NULL)
}
\arguments{
\item{parents}{[value]}

\item{checks}{[value]. Default is checks=NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
