% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFunctions.R
\name{fillPipeline}
\alias{fillPipeline}
\title{fillPipeline function}
\usage{
fillPipeline(founders, bsp = NULL, SP)
}
\arguments{
\item{founders}{Pop-class object of the founders of the breeding program}

\item{bsp}{A list of product pipeline parameters.  It contains nParents, nCrosses, nProgeny, checks, nStages, errVars, nReps, nEntries, nChks, stageNames, and nCyclesToKeepRecords.  See \code{runBreedingScheme} for details}
}
\value{
A \code{records} object. A list of lists containing nStages+1 lists. The first list contains one Pop-class of progeny per year of the scheme. The remaining lists contain one matrix per year that has individual id, mother, father, stage, phenotypes, and error variances. The individuals have been phenotyped using \code{setPheno}. The matrix may contain a mix of experimental and check phenotypes with different levels of replication
}
\description{
function to create initial records at the start of a simulation
}
\details{
This is a structure for a records object that will be used to simulate breeding schemes
}
\examples{
bsp <- specifyPipeline()
bsp <- specifyPopulation(bsp)
founderHap <- runMacs(nInd=bsp$nFounders, nChr=bsp$nChr, segSites=bsp$segSites)
SP <- SimParam$new(founderHap)
SP$addTraitA(nQtlPerChr=bsp$nQTL, var=bsp$genVar)
SP$addSnpChip(bsp$nSNP)
founders <- newPop(founderHap, simParam=SP)
bsp <- c(bsp, checks=list(NULL))
records <- fillPipeline(founders, bsp, SP)

}
