% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeByLOESS.R
\name{getParmsResponse}
\alias{getParmsResponse}
\title{getParmsResponse function}
\usage{
getParmsResponse(oneSim, startCycle, wgtPopImprov = 1)
}
\arguments{
\item{oneSim}{Output of runBreedingScheme}

\item{startCycle}{Integer first cycle after burn-in is done}

\item{wgtPopImprov}{Two components contribute to the utility: yearly population improvement and yearly gain in the product development. This is the weight for the first. The other has weight 1 - wgtPopImprov}
}
\description{
Utility function for optimizeByLOESS
Pull interesting parameters from from the stageOutputs that will serve to calculate optimality of the simulation. Currently the mean and genetic std dev of the F1 population
}
\details{
Takes a simulation output and returns a vector
}
