% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{iidPhenoEval}
\alias{iidPhenoEval}
\title{iidPhenoEval function}
\usage{
iidPhenoEval(phenoDF)
}
\arguments{
\item{phenoDF}{A data.frame of phenotypic observations. See \code{framePhenoRec} for details.}
}
\value{
Named real vector of the BLUPs of all individuals in phenoDF (names are the individual ids), with appropriate weights by error variance of the observation
}
\description{
function to take a data.frame coming from framePhenoRec and analyze it with individuals as a random effect with an IID covariance matrix
}
\details{
Given all the phenotypic records calculate the best prediction of the genotypic value for each individual using all its records
}
\examples{
phenoDF <- framePhenoRec(records, bsp)
iidBLUPs <- iidPhenoEval(phenoDF)

}
