% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFunctions.R
\name{initFuncADChk}
\alias{initFuncADChk}
\title{initFuncADChk function}
\usage{
initFuncADChk(bsp)
}
\arguments{
\item{bsp}{A list of breeding scheme parameters}
}
\value{
The return of \code{initializeScheme}
}
\description{
Deprecated function to initialize simulation of a breeding program
}
\details{
Deprecated in favor of the simply-named \code{initializeScheme}
}
