% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{populationImprovement}
\alias{populationImprovement}
\title{populationImprovement function}
\usage{
populationImprovement(records)
}
\arguments{
\item{records}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
