% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{prodPipeFncChk}
\alias{prodPipeFncChk}
\title{prodPipeFncChk function}
\usage{
prodPipeFncChk(records, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object}

\item{bsp}{A list of breeding scheme parameters}

\item{SP}{the AlphaSimR SimParam object}
}
\value{
The return from \code{productPipeline}
}
\description{
prodPipeFncChk function
}
\details{
Function deprecated in favor of the simply named \code{productPipeline}
}
