% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyingFunctions.R
\name{removeOldestCyc}
\alias{removeOldestCyc}
\title{removeOldestCyc function}
\usage{
removeOldestCyc(records, bsp)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{bsp}{The breeding scheme parameter list}
}
\value{
A \code{records} object with the first population of each list removed.
}
\description{
function to remove records of the oldest cycles still in \code{records}. Useful to avoid accumulating too much data which slows simulations down and makes them bulky
}
\details{
\code{records} is a list of lists. This function deletes the first object of each list, excluding the F1 list.
}
\examples{
records <- removeOldestCyc(records, bsp)

}
