% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{selCritGRM}
\alias{selCritGRM}
\title{selCritGRM function}
\usage{
selCritGRM(records, candidates, bsp, SP)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{candidates}{Character vector of ids of the candidates to be parents}

\item{bsp}{The breeding scheme parameter list}

\item{SP}{The AlphaSimR SimParam object (needed to pull SNPs)}
}
\value{
Character vector of the ids of the selected individuals
}
\description{
function to select parents among individuals with phenotypes, assuming individual effects covary according to a GRM
}
\details{
Accesses all individuals in \code{records} to pick the highest ones
}
\examples{
candidates <- records[[1]][[1]]@id
parents <- records[[1]][[1]][selCritGRM(records, candidates, bsp, SP)]

}
