% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPipeline.R
\name{selectAdvIID}
\alias{selectAdvIID}
\title{selectAdvIID function}
\usage{
selectAdvIID(popPheno, records, ids, bsp)
}
\arguments{
\item{popPheno}{Matrix of phenotypes of the population being selected. This matrix is provided by \code{selectInd} but need not be used}

\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{ids}{Character vector of the ids of the individuals being selected}

\item{bsp}{A list of breeding scheme parameters.}
}
\value{
Real vector of the selection criterion to chose which individuals to advance. \code{selectInd} will sort this vector and choose the individuals
}
\description{
function to select individuals to advance in the product pipeline
}
\details{
In deciding which individuals to advance, this function enables you to use all the breeding records rather than just the phenotypes of the population being selected. \code{selectInd} calls this function
}
\examples{
stage <- 2
entries <- with(bsp, {
sourcePop <- last(records[[stage]])
entries <- selectInd(sourcePop, nInd=nEntries[stage], trait=selectAdvIID, records=records, ids=sourcePop@id, bsp=bsp, simParam=SP)
entries
})
 
}
