% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprovement.R
\name{selectParents}
\alias{selectParents}
\title{selectParents function}
\usage{
selectParents(records, ppp)
}
\arguments{
\item{records}{The breeding program \code{records} object. See \code{fillPipeline} for details}

\item{ppp}{A list of product pipeline parameters.  It contains nParents, nCrosses, nProgeny, checks, nStages, errVars, nReps, nEntries, nChks, trialTypeNames, and nCyclesToKeepRecords.  See \code{runBreedingScheme} for details}
}
\value{
Character vector of the ids of the selected individuals
}
\description{
function to (do something)
}
\details{
Accesses all individuals in \code{records} to pick the highest ones
}
\examples{
none
}
