% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{specifyPipeline}
\alias{specifyPipeline}
\title{specifyPipeline function}
\usage{
specifyPipeline(bsp = NULL, big = FALSE)
}
\arguments{
\item{bsp}{A list of objects to combine with the pipeline parameters. bsp is short for breeding sheme parameters}

\item{big}{Logical TRUE for big population FALSE for toy for debugging}
}
\value{
A list containing objects that specify the product pipeline. This list will determine the number of lists in the records object
}
\description{
function specify the product pipeline. This would not need to be a function, but this way all definitions are in one place
}
\details{
Call this function before beginning the simulation
}
\examples{
bsp <- specifyPipeline()

}
