% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyPipeline.R
\name{specifyPopulation}
\alias{specifyPopulation}
\title{specifyPopulation function}
\usage{
specifyPopulation(bsp = NULL, big = FALSE)
}
\arguments{
\item{bsp}{A list of objects to combine with the species and population parameters. bsp is short for breeding sheme parameters}

\item{big}{Logical TRUE for big population FALSE for toy for debugging}
}
\value{
A list containing objects that specify the species and population characteristics.
}
\description{
Function to specify the species and population characteristics. This would not need to be a function, but this way all definitions are in one place
}
\details{
Call this function before beginning the simulation
}
\examples{
bsp <- specifyPopulation(bsp)

}
